Add in a search capability to the API
This was done in 5 minutes.  Again, Piston rocks.

One thing I do need to do is revisit the current search form.  Right now this
is different than the normal HTML form in the URLs that it uses.  Ideally,
you'd be able to just add .xml or .json to the search from the UI and get a
the raw data back as well.