Move the API url definitions to happen first.
I'm changing around the search URL structure to use JS and/or redirection
to make the search URLs match the same URI structure as the rest of the
site (i.e., add .xml or .json to get programatic info out of it).  This
reorganization is just a precursor to that.